+++ frontend/src/components/icons/google.tsx
@@ -1,0 +1,13 @@
+import { SVGProps } from 'react'
+const GoogleIcon = (props: SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 640 640"
+    width="1em"
+    height="1em"
+    {...props}
+  >
+    <path d="M564 325.8C564 467.3 467.1 568 324 568 186.8 568 76 457.2 76 320S186.8 72 324 72c66.8 0 123 24.5 166.3 64.9l-67.5 64.9c-88.3-85.2-252.5-21.2-252.5 118.2 0 86.5 69.1 156.6 153.7 156.6 98.2 0 135-70.4 140.8-106.9H324v-85.3h236.1c2.3 12.7 3.9 24.9 3.9 41.4z" />
+  </svg>
+)
+export default GoogleIcon